requirements.txt
bidict==0.22.1
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
Flask==2.2.5
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
future==1.0.0
grovepi==1.0.4
h11==0.14.0
idna==3.6
importlib-metadata==6.7.0
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
keyboard==0.13.5
MarkupSafe==2.1.5
numpy==1.21.6
pandas==1.3.5
picamera==1.13
Pillow==9.5.0
pkg-resources==0.0.0
pycparser==2.21
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-engineio==4.9.0
python-socketio==5.11.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
RPi.GPIO==0.7.1
scipy==1.7.3
simple-websocket==1.0.0
six==1.16.0
smbus-cffi==0.5.1
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
